Remove BepInEx4 compatibility Closes #42
