+++ ConfigurationManager/ConfigurationManager.csproj
@@ -49,22 +49,13 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ConfigSettingEntry.cs" />
-    <Compile Include="Obsolete\LegacySettingSearcher.cs" />
-    <Compile Include="Obsolete\Interface\AcceptableValueBaseAttribute.cs" />
-    <Compile Include="Obsolete\Interface\AcceptableValueListAttribute.cs" />
-    <Compile Include="Obsolete\Interface\AcceptableValueRangeAttribute.cs" />
-    <Compile Include="Obsolete\Interface\AdvancedAttribute.cs" />
-    <Compile Include="Obsolete\Interface\CustomSettingDrawAttribute.cs" />
+    <Compile Include="PropertySettingEntry.cs" />
     <Compile Include="SettingSearcher.cs" />
     <Compile Include="Utilities\ComboBox.cs" />
     <Compile Include="ConfigurationManager.cs" />
-    <Compile Include="Obsolete\Interface\KeyboardShortcut.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Obsolete\Interface\SavedKeyboardShortcut.cs" />
-    <Compile Include="Obsolete\LegacySettingEntry.cs" />
     <Compile Include="SettingEntryBase.cs" />
     <Compile Include="SettingFieldDrawer.cs" />
-    <Compile Include="Obsolete\FieldToPropertyInfoWrapper.cs" />
     <Compile Include="Utilities\Utilities.cs" />
     <Compile Include="ValueChangedEventArgs.cs" />
   </ItemGroup>
