--- ConfigurationManager/Obsolete/Interface/AcceptableValueBaseAttribute.cs
@@ -1,12 +1,0 @@
-﻿// Made by MarC0 / ManlyMarco
-// Copyright 2018 GNU General Public License v3.0
-
-using System;
-
-namespace BepInEx
-{
-	[Obsolete("Use the one in BepInEx.Configuration namespace")]
-    public class AcceptableValueBaseAttribute : Attribute
-	{
-    }
-}
